Don't discriminate against files starting with '-'
So far we have been pretending that any command line arg starting with a
'-' is supposed to be a switch/option. Input files with names beginning
with '-' would require being prefixed with "./" or the "--" switch. In
the case of single letter option combining this would probably make
sense, but that's not being done here and I have no plans of
implementing it. So, all the people perverted enough to begin file names
with a dash can save a couple of keystrokes from now on, I guess.